--- 07-http_gordle/final/api/guess.go
@@ -1,6 +1,0 @@
-package api
-
-// GuessRequest is the structure of the message used when submitting a guess.
-type GuessRequest struct {
-	Guess string `json:"guess"`
-}
