+++ 07-http_gordle/final/api/http.go
@@ -1,5 +1,22 @@
 package api
 
+const (
+	// GameID is the name of the field that stores the game's identifier
+	GameID = "id"
+
+	// NewGamePath is the path to create a new game.
+	NewGamePath = "/games"
+	// GetStatusPath is the path to get the status of a game identified by its id.
+	GetStatusPath = "/games/{" + GameID + "}"
+	// GuessPath is the path to play a guess in a game, identified by its id.
+	GuessPath = "/games/{" + GameID + "}"
+)
+
+// GuessRequest is the structure of the message used when submitting a guess.
+type GuessRequest struct {
+	Guess string `json:"guess"`
+}
+
 // GameResponse contains the information about a game.
 type GameResponse struct {
 	// ID is the identified of a game.
